backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
bcrypt==4.2.1
pyjwt==2.10.1
python-multipart==0.0.17
pydantic[email]==2.10.0
Pillow==11.1.0
python-dotenv==1.0.1
slowapi==0.1.9
pydantic-settings==2.7.0
pywebpush==2.0.1
